+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added support for untyped nodes. [microsoft/kiota#4385](https://github.com/microsoft/kiota/issues/4385)
+
+### Changed
+
 ## [1.3.2] - 2024-03-25
 
 ### Added